requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langgraph>=0.0.15
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
spacy>=3.5.3
httpx>=0.24.1
pydantic>=1.10.8
numpy>=1.24.3
pandas>=2.0.2
networkx>=3.1
torch>=2.0.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
